sorting parking and rental files, fetch helping function moved to utils
