+++ client/src/utils/fetchWitchCheck.js
@@ -1,0 +1,7 @@
+export default async function fetchWithCheck(url, options) {
+  const response = await fetch(url, options);
+  if (!response.ok) {
+    throw new Error('Something went wrong');
+  }
+  return response.json();
+}
